kaist-ai-functions/requirements.txt
azure-functions
langchain-google-genai
python-dotenv
google-generativeai
